Fix hidden directory check always looks at the topmost component
Before this commit, only top-level hidden directories would be skipped, so e.g. this would be found:

 → src/.hidden/foo.zig
FIXME: I'm hidden

After this commit, the basename (e.g. `.hidden`) of each entry is checked (instead of the path, e.g. `src/.hidden`), so hidden directories that aren't at the top-level of the tree are skipped as well.